requirements.txt
# requirements.txt
streamlit==1.31.0
firebase-admin==6.4.0
langchain==0.1.0
openai==1.12.0
google-cloud-firestore==2.15.0
python-dotenv==1.0.0
